backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0

# CORS handling for React
django-cors-headers==4.3.0

# Image processing
# Pillow==10.2.0

# Rich text editor
django-ckeditor==6.7.0

# Automatic cleanup of old files
django-cleanup==8.0.0

# Environment variables management
python-decouple==3.8

# Database (PostgreSQL support - optional)
# psycopg2-binary==2.9.9

# Timezone support
pytz==2023.3

# Security
django-environ==0.11.2

# API documentation (optional but recommended)
drf-yasg==1.21.7

# Filtering for API
django-filter==23.5

# WSGI server for production
gunicorn==21.2.0

# Static files management for production
whitenoise==6.6.0

# Email backend (optional)
django-anymail==10.2

# Celery for async tasks (optional)
celery==5.3.4
redis==5.0.1

# Development tools
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code quality
flake8==6.1.0
black==23.12.1
isort==5.13.2